requirements.txt
pydantic
langchain
langgraph
pillow
fastapi
requests
beautifulsoup4
reportlab
uvicorn[standard]
python-dotenv
langchain-community 
httpx
langchain-ollama
pvporcupine
pyaudio
faster-whisper
numpy
psycopg2-binary
sqlalchemy